% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_generator.R
\name{generate_numbers}
\alias{generate_numbers}
\title{Generate random numbers.}
\usage{
generate_numbers(sample_size, distribution, ...)
}
\arguments{
\item{sample_size}{A number indicating how many random numbers to return.}

\item{distribution}{Either "normal", "poisson" or "binomial" to indicate distribution type.}

\item{...}{Other arguments passed on.}
}
\value{
A vector of random numbers distributed as specified.
}
\description{
Generate random numbers.
}
\examples{
generate_numbers(100, "normal")
generate_numbers(100, "poisson", lambda = 10)
}
