% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_generator.R
\name{summary.numbers_class}
\alias{summary.numbers_class}
\title{Summary function for numbers_class objects.}
\usage{
\method{summary}{numbers_class}(object, ...)
}
\arguments{
\item{object}{A numbers_class object.}

\item{...}{Additional arguments affecting the summary produced - not in use.}
}
\value{
A summary of the object, including min, max, median, mean and IQR.
}
\description{
Summary function for numbers_class objects.
}
\examples{
summary(generate_numbers(100, "normal"))
}
